Replacing Ox with REXML
This replaces the Ox XML library with REXML from Ruby's standard
library.

The files being parsed by Ow are relatively small and does not require
stream parsing, so the performance improvement offered by Ox is not
worth the cost of having an external dependency.

Also, Ox is a C extension. Removing it means Ow can run on JRuby.

Signed-off-by: Martin Svangren <martin@masv.net>